ENG-11789, improving mesh arbitor to shorten arbitration time.
When one host starts to send out its final decision and waits for other hosts to agree with it,
it means this host have the enough information to decide which host is dead. So if it receives
a message but not carry with new information (no new failed node or new survivor), there is no
need to interrupt current loop of arbitration and to broadcast the message to other hosts.

This patch addresses the above problem, makes arbitration process shorter and easier to reach
an agreement.

Change-Id: Ife3f6785ce4fe86e552813963fb90bd1915d6582